test(P10-F06): cover Active/Historic summary services, facade dispatch, and realized totals
Adapts the pre-split summary tests into ActivePortfolioAssetSummaryServiceTests,
adds HistoricPortfolioAssetSummaryServiceTests covering gross-TotalBought
weighting and RealizedGainLoss (including the credit-percentage fields
no longer collapsing for a closed position), adds PortfolioAssetSummaryServiceTests
for facade dispatch, and extends the scope-historic integration test with
value assertions plus an Active-scope regression test.
